added brain interface
